Add intrinsic function

+++ ast.hpp
@@ -22,10 +22,10 @@ enum OpType {
 };
 
 class ExprAST {
-  long long pos;
+  int pos;
 public:
-  ExprAST(long long pos) : pos(pos) {}
-  long long getPos() { return pos; }
+  ExprAST(int pos) : pos(pos) {}
+  int getPos() { return pos; }
   virtual ~ExprAST() {}
   virtual Value* codegen() = 0;
   virtual void write(std::ostream& os) = 0;
@@ -34,7 +34,7 @@ public:
 class ArgExprAST : public ExprAST {
   int n;
 public:
-  ArgExprAST(long long pos, int n) : ExprAST(pos), n(n) {}
+  ArgExprAST(int pos, int n) : ExprAST(pos), n(n) {}
   Value* codegen() override;
   void write(std::ostream& os) override;
 };
@@ -42,7 +42,7 @@ public:
 class MutableVarExprAST : public ExprAST {
   std::string name;
 public:
-  MutableVarExprAST(long long pos, std::string name) : ExprAST(pos), name(name) {}
+  MutableVarExprAST(int pos, std::string name) : ExprAST(pos), name(name) {}
   std::string& getName() { return name; }
   Value* codegen() override;
   void write(std::ostream& os) override;
@@ -51,7 +51,7 @@ public:
 class IntExprAST : public ExprAST {
   int64_t val;
 public:
-  IntExprAST(long long pos, int64_t val) : ExprAST(pos), val(val) {}
+  IntExprAST(int pos, int64_t val) : ExprAST(pos), val(val) {}
   Value* codegen() override;
   void write(std::ostream& os) override;
 };
@@ -60,7 +60,7 @@ class BinaryOpExprAST : public ExprAST {
   OpType op;
   std::unique_ptr<ExprAST> lhs, rhs;
 public:
-  BinaryOpExprAST(long long pos, OpType op, std::unique_ptr<ExprAST> lhs, std::unique_ptr<ExprAST> rhs) 
+  BinaryOpExprAST(int pos, OpType op, std::unique_ptr<ExprAST> lhs, std::unique_ptr<ExprAST> rhs) 
     : ExprAST(pos), op(op), lhs(std::move(lhs)), rhs(std::move(rhs)) {}
   Value* codegen() override;
   void write(std::ostream& os) override;
@@ -69,7 +69,7 @@ public:
 class BoolExprAST : public ExprAST {
   bool val;
 public:
-  BoolExprAST(long long pos, bool val) : ExprAST(pos), val(val) {}
+  BoolExprAST(int pos, bool val) : ExprAST(pos), val(val) {}
   Value* codegen() override;
   void write(std::ostream& os) override;
 };
@@ -77,7 +77,7 @@ public:
 class IfExprAST : public ExprAST {
   std::unique_ptr<ExprAST> cnd, thn, els;
 public:
-  IfExprAST(long long pos, std::unique_ptr<ExprAST> cnd, std::unique_ptr<ExprAST> thn, std::unique_ptr<ExprAST> els)
+  IfExprAST(int pos, std::unique_ptr<ExprAST> cnd, std::unique_ptr<ExprAST> thn, std::unique_ptr<ExprAST> els)
     : ExprAST(pos), cnd(std::move(cnd)), thn(std::move(thn)), els(std::move(els)) {}
   Value* codegen() override;
   void write(std::ostream& os) override;
@@ -86,7 +86,7 @@ public:
 class SeqExprAST : public ExprAST {
   std::unique_ptr<ExprAST> fst, snd;
 public:
-  SeqExprAST(long long pos, std::unique_ptr<ExprAST> fst, std::unique_ptr<ExprAST> snd)
+  SeqExprAST(int pos, std::unique_ptr<ExprAST> fst, std::unique_ptr<ExprAST> snd)
     : ExprAST(pos), fst(std::move(fst)), snd(std::move(snd)) {}
   Value* codegen() override;
   void write(std::ostream& os) override;
@@ -95,7 +95,7 @@ public:
 class WhileExprAST : public ExprAST {
   std::unique_ptr<ExprAST> cnd, body;
 public:
-  WhileExprAST(long long pos, std::unique_ptr<ExprAST> cnd, std::unique_ptr<ExprAST> body)
+  WhileExprAST(int pos, std::unique_ptr<ExprAST> cnd, std::unique_ptr<ExprAST> body)
     : ExprAST(pos), cnd(std::move(cnd)), body(std::move(body)) {}
   Value* codegen() override;
   void write(std::ostream& os) override;
@@ -104,7 +104,7 @@ public:
 class SetExprAST : public ExprAST {
   std::unique_ptr<ExprAST> val, var;
 public:
-  SetExprAST(long long pos, std::unique_ptr<ExprAST> val, std::unique_ptr<ExprAST> var)
+  SetExprAST(int pos, std::unique_ptr<ExprAST> val, std::unique_ptr<ExprAST> var)
     : ExprAST(pos), val(std::move(val)), var(std::move(var)) {}
   Value* codegen() override;
   void write(std::ostream& os) override;
